+++ Applications/WebBrowser/desc.sh
@@ -1,11 +1,12 @@
 #!/bin/sh
 
 APPNAME="WebBrowser"
-VERSION=""
+VERSION="0.1"
 SHORTNAME="webbrowser"
 REPO="gs-webbrowser"
 METHOD="git"					# git | get
 HUB="https://github.com/onflapp/"
+BRANCH="main"
 PATH_SRC=""
 FRAMEWORKS=""
 DEPS="YES"