+++ Draft/ModelApp/desc.sh
@@ -3,16 +3,18 @@
 APPNAME="AppName"
 VERSION="1.0.0"
 SHORTNAME="appname"
+ALIAS=""
 REPO="gap"
 METHOD="git"					# "git" | "get"
 HUB="https://github.com/gnustep/"
+BRANCH="master"					# "master" | "main"
 EXT=""					# if get method: "zip" | "tar.gz"...
 PATH_SRC="user-apps/${APPNAME}"
 FRAMEWORKS=""
-DEPS="YES"					# "YES" or leave empty
+DEPS=""					# "YES" or leave empty
 
 # READ SPECIFIC INFOS IN SOURCE FOLDER
-BUILD_ARGS="foo"
+BUILD_ARGS=""
 INSTALL_ARGS="${BUILD_ARGS}"
 
 PATCH=""			# path of the file.patch or leave empty