+++ Frameworks/Pantomime/3_install.sh
@@ -24,4 +24,7 @@ sudo -E make install ${INSTALL_ARGS}
 
 make clean
 
+# Fixing Linker issue
+sudo ldconfig
+
 cd ${_PWD}